updated tests for new ccmp-client.jar #26372

+++ library-test/src/androidTest/java/net/ut11/ccmp/lib/JsonTest.java
@@ -38,7 +38,7 @@ public class JsonTest extends TestCase {
 		DeviceResponse resp = JSONUnmarshaller.unmarshall(jsonString, DeviceResponse.class);
 
 		assertEquals(resp.getApiKey(), "123456abcdef#?-/\"");
-		assertEquals(resp.getMsisdn(), 436761234567L);
+		assertEquals((long) resp.getMsisdn(), 436761234567L);
 		assertFalse(resp.getEnabled());
 	}
